chore : 설정파일 및 의존성 수정
