+++ api-gateway-server/src/main/resources/application.yml
@@ -1,0 +1,78 @@
+server:
+  port: 6000
+
+spring:
+  application:
+    name: api-gateway-server
+  profiles:
+    default: local
+  cloud:
+    compatibility-verifier:
+      enabled: false
+    gateway:
+      routes:
+      - id: authentication-server
+        uri: lb://authentication-service
+        predicates:
+          - Path=/auth/**
+        filters:
+          - RewritePath=/auth/(?<segment>.*), /api/v1/auth/${segment}
+      - id: chat-server
+        uri: lb://chat-service
+        predicates:
+          - Path=/chat/**
+        filters:
+          - RewritePath=/chat/(?<segment>.*), /api/v1/chat/${segment}
+      - id: cdn-server
+        uri: lb://cdn-service
+        predicates:
+          - Path=/cdn/**
+        filters:
+          - RewritePath=/cdn/(?<segment>.*), /api/v1/cdn/${segment}
+      - id: game-server
+        uri: lb://game-service
+        predicates:
+          - Path=/game/**
+        filters:
+          - RewritePath=/game/(?<segment>.*), /api/v1/game/${segment}
+      - id: api-service-server
+        uri: lb://api-service
+        predicates:
+          - Path=/api/**
+        filters:
+          - RewritePath=/api/(?<segment>.*), /api/v1/api/${segment}
+      - id: logging-server
+        uri: lb://logging-service
+        predicates:
+          - Path=/logging/**
+        filters:
+          - RewritePath=/logging/(?<segment>.*), /api/v1/logging/${segment}
+      - id: matching-server
+        uri: lb://matching-service
+        predicates:
+          - Path=/matching/**
+        filters:
+          - RewritePath=/matching/(?<segment>.*), /api/v1/matching/${segment}
+  servlet:
+    multipart:
+      max-file-size: 5MB
+      max-request-size: 5MB
+
+  autoconfigure:
+    exclude:
+      - org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration
+      - org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
+
+management:
+  endpoints:
+    web:
+      base-path: /healthcheck
+      exposure:
+        include: "*"
+
+logging:
+  level:
+    org:
+      hibernate:
+        type: info
+        level: info
