+++ api-service-server/build.gradle
@@ -1,31 +1,30 @@
 dependencies {
     //common-module
     implementation project(':common-module')
+    annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
     //Web
     implementation 'org.springframework.boot:spring-boot-starter-web'
     //Test
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    //JPA
+    // JPA
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    //PostgreSQL
+	// PostgreSQL
     implementation 'org.postgresql:postgresql'
     //Redis
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+
     //Security
     implementation "org.springframework.security:spring-security-crypto"
-	//JWT
-    implementation 'io.jsonwebtoken:jjwt-api:0.12.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.5'
+
     //QueryDSL
-    implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
-    annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jakarta"
-    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
-    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+	implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
+	annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jakarta"
+	annotationProcessor "jakarta.annotation:jakarta.annotation-api"
+	annotationProcessor "jakarta.persistence:jakarta.persistence-api"
 }
 
 clean {
-    delete file('src/main/generated')
+	delete file('src/main/generated')
 }
 
 test {
