+++ api-service-server/src/main/java/com/lolclone/api_service_server/DatabaseServerApplication.java
@@ -1,0 +1,13 @@
+package com.lolclone.api_service_server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DatabaseServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DatabaseServerApplication.class, args);
+	}
+
+}
