+++ api-service-server/src/main/resources/application.yml
@@ -1,0 +1,18 @@
+server:
+    port: 6005
+
+spring:
+    profiles:
+        default: local
+    
+    application:
+        name: api-service-server
+    
+    config:
+        import:
+            - security/application-db.yml
+    
+    servlet:
+        multipart:
+            max-file-size: 5MB
+            max-request-size: 5MB
