+++ api-service-server/src/test/java/com/lolclone/api_service_server/DatabaseServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lolclone.api_service_server;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DatabaseServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
