+++ build.gradle
@@ -30,13 +30,9 @@ subprojects {
     dependencies {
         //Validation
 	    implementation 'org.springframework.boot:spring-boot-starter-validation'
-        //Web
-        implementation 'org.springframework.boot:spring-boot-starter-web'
         //Lombok
         compileOnly 'org.projectlombok:lombok'
         annotationProcessor 'org.projectlombok:lombok'
-        //Test
-        testImplementation 'org.springframework.boot:spring-boot-starter-test'
         testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
         testImplementation 'org.projectlombok:lombok'
         testAnnotationProcessor 'org.projectlombok:lombok'
