+++ chat-server/src/main/resources/application.yml
@@ -1,0 +1,18 @@
+server:
+    port: 6002
+
+spring:
+    profiles:
+        default: local
+    
+    application:
+        name: chat-server
+    
+    config:
+        import:
+            - security/application-db.yml
+    
+    servlet:
+        multipart:
+            max-file-size: 5MB
+            max-request-size: 5MB
