+++ common-module/src/main/resources/application.yml
@@ -1,0 +1,1 @@
+
