+++ content-delivery-server/build.gradle
@@ -1,4 +1,17 @@
 dependencies {
 	//common-module
     implementation project(':common-module')
+    //Web
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    //Test
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    //JPA
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    //PostgreSQL
+    implementation 'org.postgresql:postgresql'
+    //QueryDSL
+    implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
+    annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jakarta"
+    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
+    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
 }
