--- database-server/src/main/java/com/lolclone/database_server/DatabaseServerApplication.java
@@ -1,13 +1,0 @@
-package com.lolclone.database_server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DatabaseServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DatabaseServerApplication.class, args);
-	}
-
-}
