--- database-server/src/main/java/com/lolclone/database_server/authenticationserver/domain/RegistrationType.java
@@ -1,5 +1,0 @@
-package com.lolclone.database_server.authenticationserver.domain;
-
-public enum RegistrationType {
-    STANDARD, SOCIAL
-}
