--- database-server/src/main/java/com/lolclone/database_server/authenticationserver/domain/SocialType.java
@@ -1,11 +1,0 @@
-package com.lolclone.database_server.authenticationserver.domain;
-
-public enum SocialType {
-    FACEBOOK,
-    GOOGLE,
-    KAKAO,
-    XBOX,
-    PLAYSTATION,
-    APPLE,
-    NONE
-}
