--- database-server/src/main/java/com/lolclone/database_server/authenticationserver/domain/UserInfo.java
@@ -1,13 +1,0 @@
-package com.lolclone.database_server.authenticationserver.domain;
-
-import lombok.Builder;
-
-@Builder
-public record UserInfo(
-    String socialId,
-    SocialType socialType,
-    String nickname,
-    String profileImage
-) {
-    
-}
