--- database-server/src/main/java/com/lolclone/database_server/authenticationserver/domain/UserRole.java
@@ -1,6 +1,0 @@
-package com.lolclone.database_server.authenticationserver.domain;
-
-public enum UserRole {
-    ROLE_USER,
-    ROLE_ANONYMOUS
-}
