--- database-server/src/main/java/com/lolclone/database_server/authenticationserver/domain/UserStatus.java
@@ -1,9 +1,0 @@
-package com.lolclone.database_server.authenticationserver.domain;
-
-public enum UserStatus {
-    ONLINE,
-    OFFLINE,
-    IN_GAME,
-    CUSTOM_GAME,
-    MATCHMAKING,
-}
