--- database-server/src/main/java/com/lolclone/database_server/config/JpaAuditingConfig.java
@@ -1,10 +1,0 @@
-package com.lolclone.database_server.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-    
-}
