--- database-server/src/main/resources/application.yml
@@ -1,18 +1,0 @@
-server:
-    port: 6004
-
-spring:
-    profiles:
-        default: local
-    
-    application:
-        name: database-server
-    
-    config:
-        import:
-            - security/application-db.yml
-    
-    servlet:
-        multipart:
-            max-file-size: 5MB
-            max-request-size: 5MB
