--- database-server/src/test/java/com/lolclone/database_server/DatabaseServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.lolclone.database_server;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DatabaseServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
