+++ settings.gradle
@@ -4,7 +4,7 @@ include('game-server')
 include('matchmaking-server')
 include('authentication-management-server')
 include('chat-server')
-include('database-server')
+include('api-service-server')
 include('api-gateway-server')
 include('logging-monitoring-server')
 include('content-delivery-server')
